+++ scripts/preblast_reminders.py
@@ -132,7 +132,9 @@ def send_preblast_reminders(force: bool = False):
                     EventInstance.preblast_rich.is_(None), EventInstance.preblast_rich.cast(String) == "null"
                 ),  # not already set
                 EventInstance.is_active,  # not canceled
-                EventInstance.series_exception != Series_Exception.closed,  # not closed
+                or_(
+                    EventInstance.series_exception.is_(None), EventInstance.series_exception != Series_Exception.closed
+                ),  # noqa: E501
             ]
         )
         preblast_list.items = [item for item in preblast_list.items if item.q_name is not None]